+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 
 3.2.1
 
-* API update: `HashIndex::peek` no longer allows entry references to outlive the `HashIndex`: [#198](https://github.com/wvwwvwwv/scalable-concurrent-containers/issues/198).
+* API update: `HashIndex::{peek, iter}` no longer allows entry references to outlive the `HashIndex`: [#198](https://github.com/wvwwvwwv/scalable-concurrent-containers/issues/198).
 * More memory usage optimization: [#194](https://github.com/wvwwvwwv/scalable-concurrent-containers/issues/194).
 
 3.2.0